+++ src/main/kotlin/id/walt/vclib/Defaults.kt
@@ -1,9 +1,10 @@
 package id.walt.vclib
 
-import id.walt.vclib.vclist.*
+import id.walt.vclib.credentials.*
 //ANDROID PORT
 //import org.lighthousegames.logging.logging
 //ANDROID PORT
+
 
 object Defaults {