+++ src/test/kotlin/id/walt/vclib/schema/SchemaServiceTest.kt
@@ -1,19 +1,18 @@
 package id.walt.vclib.schema
 
 import com.beust.klaxon.Json
+import id.walt.vclib.credentials.*
 import id.walt.vclib.model.VerifiableCredential
 import id.walt.vclib.registry.VerifiableCredentialMetadata
-import io.kotest.assertions.json.shouldEqualJson
+import id.walt.vclib.schema.SchemaService.DateTimeFormat
+import id.walt.vclib.schema.SchemaService.JsonIgnore
+import id.walt.vclib.schema.SchemaService.PropertyName
 import io.kotest.core.spec.style.StringSpec
 import java.io.File
-import id.walt.vclib.schema.SchemaService.DateTimeFormat;
-import id.walt.vclib.schema.SchemaService.JsonIgnore;
-import id.walt.vclib.schema.SchemaService.Nullable;
-import id.walt.vclib.schema.SchemaService.PropertyName;
 
 data class DummyCredential(
-    @Json(name = "@context") @field:PropertyName(name = "@context") var context: List<String> = listOf("https://www.w3.org/2018/credentials/v1"),
-    @Json(serializeNull = false) @field:Nullable override var id: String? = null,
+    @Json(name = "@context") @field:PropertyName(name = "@context") @field:SchemaService.Required var context: List<String> = listOf("https://www.w3.org/2018/credentials/v1"),
+    @Json(serializeNull = false) @field:SchemaService.Required override var id: String? = null,
     @Json(serializeNull = false) @field:DateTimeFormat var issuanceDate: String? = null,
     @field:JsonIgnore var toIgnore: String? = null
 ) : VerifiableCredential(type) {
@@ -26,8 +25,47 @@ data class DummyCredential(
 class SchemaServiceTest : StringSpec({
 
     "testing schema generation"   {
-        val schema = SchemaService.generateSchema(DummyCredential::class).toPrettyString()
-        val expected = File("src/test/resources/schemas/DummyCredential-schema.json").readText()
-        schema shouldEqualJson expected
+        generateSchema(DummyCredential::class.java)
+    }
+
+    "verify VerifiableDiploma schema"   {
+        generateSchema(VerifiableDiploma::class.java)
+    }
+
+    "verify VerifiableAttestation schema"   {
+        generateSchema(VerifiableAttestation::class.java)
+    }
+
+    "verify VerifiableAuthorization schema"   {
+        generateSchema(VerifiableAuthorization::class.java)
+    }
+
+    "verify VerifiablePresentation schema"   {
+        generateSchema(VerifiablePresentation::class.java)
+    }
+
+    "verify VerifiableId schema"   {
+        generateSchema(VerifiableId::class.java)
+    }
+
+    "verify Europass schema"   {
+        generateSchema(Europass::class.java)
+    }
+
+    "verify GaiaxCredential schema"   {
+        generateSchema(GaiaxCredential::class.java)
+    }
+
+    "verify UniversityDegree schema"   {
+        generateSchema(UniversityDegree::class.java)
+    }
+
+    "verify PermanentResidentCard schema"   {
+        generateSchema(PermanentResidentCard::class.java)
     }
 })
+
+private inline fun <T : VerifiableCredential> generateSchema(vc: Class<T>) {
+    val schema = SchemaService.generateSchema(vc)
+    File("src/test/resources/schemas/${vc.simpleName}.json").writeText(schema)
+}